requirements.txt
pandas
requests
python-dotenv
beautifulsoup4
lxml
kaggle
matplotlib
seaborn